+++ backend/src/main/java/org/camunda/optimize/service/es/report/result/NumberResult.java
@@ -7,5 +7,5 @@ package org.camunda.optimize.service.es.report.result;
 
 public interface NumberResult {
 
-  long getResultAsNumber();
+  double getResultAsNumber();
 }
